Выполнил домашнее задание по личному проекту
