+++ src/components/trip-list.js
@@ -1,0 +1,7 @@
+const createTripListTemplate = () => {
+  return (
+    `<ul class="trip-days"></ul>`
+  );
+};
+
+export {createTripListTemplate};
